refactor: move roslyn_handlers to seperate module
